Split the BlobService in a generic lib and a SqlServer implementation lib. Update namespaces and nullables in the process.
