+++ Core/Blob.Data.SqlServer/Core.Blob.Data.SqlServer/BlobDbContext.cs
@@ -1,0 +1,27 @@
+﻿using Lens.Core.Blob.Data.Entities;
+using Lens.Core.Data.EF;
+using Lens.Core.Data.EF.Services;
+using Lens.Core.Data.Services;
+using Lens.Core.Lib.Services;
+using Microsoft.EntityFrameworkCore;
+
+namespace Lens.Core.Blob.Data;
+
+public class BlobDbContext : ApplicationDbContext
+{
+    public BlobDbContext(DbContextOptions<BlobDbContext> options,
+        IUserContext userContext,
+        IAuditTrailService auditTrailService,
+        IEnumerable<IModelBuilderService> modelBuilders) : base(options, userContext, auditTrailService, modelBuilders)
+    {
+    }
+
+    public DbSet<BlobInfo> BlobInfos { get; set; } = null!;
+
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+        base.OnModelCreating(modelBuilder);
+
+        ConfigureBaseProperties(typeof(BlobInfo), modelBuilder);
+    }
+}
