+++ Core/Blob.Data.SqlServer/Core.Blob.Data.SqlServer/Core.Blob.Data.SqlServer.csproj
@@ -1,0 +1,14 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\Blob\Core.Blob.csproj" />
+    <ProjectReference Include="..\..\Data.EF.SqlServer\Core.Data.EF.SqlServer\Core.Data.EF.SqlServer.csproj" />
+  </ItemGroup>
+
+</Project>
