+++ Core/Blob.Data.SqlServer/Core.Blob.Data.SqlServer/readme.md
