+++ Core/Blob/Models/BlobInfoInputModel.cs
@@ -1,18 +1,16 @@
-﻿using System;
-using System.ComponentModel.DataAnnotations;
+﻿using System.ComponentModel.DataAnnotations;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 
-namespace Lens.Core.Blob.Models
+namespace Lens.Core.Blob.Models;
+
+public class BlobInfoInputModel
 {
-    public class BlobInfoInputModel
-    {
-        [Required]
-        public Guid? EntityId { get; set; }
-        [StringLength(532)]
-        public bool SkipFileDeletion { get; set; } = false;
-        [JsonIgnore]
-        public string Tag => JsonSerializer.Serialize(Tags ?? Array.Empty<string>());
-        public string[] Tags { get; set; }
-    }
+    [Required]
+    public Guid? EntityId { get; set; }
+    [StringLength(532)]
+    public bool SkipFileDeletion { get; set; } = false;
+    [JsonIgnore]
+    public string Tag => JsonSerializer.Serialize(Tags ?? Array.Empty<string>());
+    public string[]? Tags { get; set; }
 }
