+++ Core/Blob/Models/FileInfoModel.cs
@@ -1,9 +1,8 @@
-﻿namespace Lens.Core.Blob.Models
+﻿namespace Lens.Core.Blob.Models;
+
+public class FileInfoModel
 {
-    public class FileInfoModel
-    {
-        public string FileExtension { get; set; }
-        public string FilenameWithoutExtension { get; set; }
-        public string RelativePathAndName { get; set; }
-    }
+    public string? FileExtension { get; set; }
+    public string? FilenameWithoutExtension { get; set; }
+    public string RelativePathAndName { get; set; } = null!;
 }
