+++ Core/Blob/Services/AzureStorageBlobService.cs
@@ -1,46 +1,42 @@
 ﻿using Lens.Core.Blob.Models;
 using Lens.Core.Lib.Services;
 using Microsoft.Extensions.Configuration;
-using System;
-using System.IO;
-using System.Threading.Tasks;
 
-namespace Lens.Core.Blob.Services
+namespace Lens.Core.Blob.Services;
+
+public class AzureStorageBlobService : BaseService<AzureStorageBlobService>, IBlobService
 {
-    public class AzureStorageBlobService : BaseService<AzureStorageBlobService>, IBlobService
+    private readonly BlobSettings _blobServiceSettings;
+
+    public AzureStorageBlobService(
+        IApplicationService<AzureStorageBlobService> applicationService,
+        IConfiguration configuration) : base(applicationService)
     {
-        private readonly BlobSettings _blobServiceSettings;
+        _blobServiceSettings = configuration.GetSection(nameof(BlobSettings)).Get<BlobSettings>();
+    }
 
-        public AzureStorageBlobService(
-            IApplicationService<AzureStorageBlobService> applicationService,
-            IConfiguration configuration) : base(applicationService)
-        {
-            _blobServiceSettings = configuration.GetSection(nameof(BlobSettings)).Get<BlobSettings>();
-        }
+    public Task<bool> DeleteBlob(string relativePathAndName)
+    {
+        throw new NotImplementedException();
+    }
 
-        public Task<bool> DeleteBlob(string relativePathAndName)
-        {
-            throw new NotImplementedException();
-        }
+    public Task<Stream> Download(string relativePathAndName)
+    {
+        throw new NotImplementedException();
+    }
 
-        public Task<Stream> Download(string relativePathAndName)
-        {
-            throw new NotImplementedException();
-        }
+    public Task<string[]> GetBlobs()
+    {
+        throw new NotImplementedException();
+    }
 
-        public Task<string[]> GetBlobs()
-        {
-            throw new NotImplementedException();
-        }
+    public Task<string> GetBlobUrl(string relativePathAndName)
+    {
+        throw new NotImplementedException();
+    }
 
-        public Task<string> GetBlobUrl(string relativePathAndName)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Task<BlobMetadataModel> Upload(string relativePathAndName, Stream stream)
-        {
-            throw new NotImplementedException();
-        }
+    public Task<BlobMetadataModel> Upload(string relativePathAndName, Stream stream)
+    {
+        throw new NotImplementedException();
     }
 }
