+++ Core/Blob/Services/IBlobService.cs
@@ -1,15 +1,12 @@
 ﻿using Lens.Core.Blob.Models;
-using System.IO;
-using System.Threading.Tasks;
 
-namespace Lens.Core.Blob.Services
+namespace Lens.Core.Blob.Services;
+
+public interface IBlobService
 {
-    public interface IBlobService
-    {
-        Task<BlobMetadataModel> Upload(string relativePathAndName, Stream stream);
-        Task<Stream> Download(string relativePathAndName);
-        Task<string[]> GetBlobs();
-        Task<string> GetBlobUrl(string relativePathAndName);
-        Task<bool> DeleteBlob(string relativePathAndName);
-    }
+    Task<BlobMetadataModel> Upload(string relativePathAndName, Stream stream);
+    Task<Stream> Download(string relativePathAndName);
+    Task<string[]> GetBlobs();
+    Task<string> GetBlobUrl(string relativePathAndName);
+    Task<bool> DeleteBlob(string relativePathAndName);
 }
